Refactor and comment the dataset generation module
