+++ src/lib/features/generate-dataset/index.ts
@@ -1,5 +1,5 @@
-export { dataset } from './model/store';
+export { dataset } from './model/dataset';
 export { addImage } from './model/add-image';
 export { removeLastImage } from './model/remove-last-image';
-export { exportData } from './model/export-data';
+export { downloadDataset } from './model/download-dataset';
 export type { Sample, SampleClass } from './model/type';
