+++ src/lib/features/generate-dataset/model/dataset.ts
@@ -2,4 +2,5 @@ import { writable } from 'svelte/store';
 
 import type { Sample } from './type';
 
+/** A global in-memory dataset. */
 export const dataset = writable<Sample[]>([]);
