+++ src/lib/features/generate-dataset/model/remove-last-image.ts
@@ -1,5 +1,6 @@
-import { dataset } from './store';
+import { dataset } from './dataset';
 
+/** Remove the last image from the global dataset. */
 export function removeLastImage(): void {
   dataset.update(($dataset) => {
     $dataset.pop();
