+++ src/lib/features/generate-dataset/model/type.d.ts
@@ -1,8 +1,14 @@
-export type SampleClass = 'Circle' | 'Rectangle' | 'Nothing';
+import type { Shape } from '$lib/shared/train-data';
+
+export type SampleClass = Shape | 'Nothing';
 
 export interface Sample {
   data: Blob;
-  /** `[[left, top], [left + width, top + height]]` */
+  /**
+   * A bounding box in XYXY format.
+   *
+   * The four numbers are `[left, top, left + width, top + height]`.
+   */
   coordinates: [[number, number], [number, number]];
   className: SampleClass;
 }
